+++ src/main/java/com/nnk/springboot/dto/UpdateCurvePointDTO.java
@@ -4,6 +4,7 @@
 package com.nnk.springboot.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.validation.constraints.Digits;
 import javax.validation.constraints.NotNull;
@@ -35,7 +36,7 @@ public class UpdateCurvePointDTO implements Serializable {
     @NotNull(message = "Term is mandatory.")
 	@Positive(message = "Term must be strictly positive.")
 	@Digits(integer = 10, fraction = 2, message = "Term must be a decimal number.")
-	private double term;
+	private BigDecimal term;
 	
 	/**
 	 * Value.
@@ -43,5 +44,5 @@ public class UpdateCurvePointDTO implements Serializable {
     @NotNull(message = "Value is mandatory.")
 	@Positive(message = "Value must be strictly positive.")
 	@Digits(integer = 10, fraction = 2, message = "Value must be a decimal number.")
-	private double value;
+	private BigDecimal value;
 }
